+++ src/InvoiceData/Service/InvoiceFileService.php
@@ -12,12 +12,12 @@ namespace FreshAdvance\Invoice\InvoiceData\Service;
 use FreshAdvance\Invoice\InvoiceData\DataType\InvoiceDataInterface;
 use FreshAdvance\Invoice\Pdf\Service\FilenameCalculatorInterface;
 use FreshAdvance\Invoice\Settings\ModuleSettingsInterface;
-use FreshAdvance\Invoice\Transput\UtilsInterface;
+use FreshAdvance\Invoice\Transput\ResponseInterface;
 
 class InvoiceFileService implements InvoiceFileServiceInterface
 {
     public function __construct(
-        protected readonly UtilsInterface $utils,
+        protected readonly ResponseInterface $utils,
         protected readonly ModuleSettingsInterface $moduleSettings,
         protected readonly FilenameCalculatorInterface $filenameCalculator,
     ) {